refactor: extract settingsApi to a helper file
